Navigator & Cauldron
Cauldron is finally finished! However, further tweaking and testing is required to say mod asset complete. Also introduced the Navigator, it's base Info is set up, along with Idle and Attack animations, However  the Rally AI needs more work on it but I'm not too focused on finishing it now. Time for the Golem!